refactor: Extract MCP tools layer (Part 3 of modularization)
Final extraction of MCP tool definitions from mcp_server.py:
- Created tools/search_tool.py with search_tool() MCP function (84 lines)
- Created tools/health_check_tool.py with health_check_tool() MCP function (23 lines)
- Updated mcp_server.py to import and register tools dynamically
- Removed duplicate code and unnecessary imports from mcp_server.py
- Simplified logging message for API key status

Impact:
- mcp_server.py reduced from 208 to 120 lines (42% reduction from Part 2)
- mcp_server.py reduced from 496 to 120 lines overall (76% total reduction)
- Clean separation: Entry point now only handles logging setup, tool registration, and server startup
- Each tool is self-contained with its own dependencies

Files changed:
- docker/tools/search_tool.py (NEW) - 84 lines
- docker/tools/health_check_tool.py (NEW) - 23 lines
- docker/mcp_server.py (MODIFIED) - 120 lines (was 208, originally 496)

Architecture now complete:
├── models/       - Pydantic data models (5 files)
├── clients/      - External API integrations (2 files)
├── services/     - Business logic (2 files)
├── tools/        - MCP tool definitions (2 files)
└── mcp_server.py - Entry point (120 lines)

Part 3 of 3: Complete! ✓
